requirements.txt
aiogram>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
